add authCtrl & refactor router
