+++ docs/swagger.json
@@ -88,12 +88,12 @@
                 "summary": "User register",
                 "parameters": [
                     {
-                        "description": "userInfoReqBody",
+                        "description": "SignupReqBody",
                         "name": "request",
                         "in": "body",
                         "required": true,
                         "schema": {
-                            "$ref": "#/definitions/api.userInfoReqBody"
+                            "$ref": "#/definitions/controller.SignupReqBody"
                         }
                     }
                 ],
@@ -109,7 +109,7 @@
         }
     },
     "definitions": {
-        "api.userInfoReqBody": {
+        "controller.SignupReqBody": {
             "type": "object",
             "properties": {
                 "checkPassword": {
