+++ router.go
@@ -1,8 +1,9 @@
 package main
 
 import (
-	api "gocard/api"
+	controllers "gocard/controllers"
 	route "gocard/route"
+	service "gocard/service"
 
 	"github.com/gin-gonic/gin"
 	swaggerFiles "github.com/swaggo/files"
@@ -16,11 +17,11 @@ func initRouter() *gin.Engine {
 	// router
 	v1 := server.Group("/dev/api/v1")
 	{
-		v1.POST("/login", api.AuthHandler)
-		v1.POST("/signup", api.RegisterHandler)
+		v1.POST("/login", controllers.AuthHandler)
+		v1.POST("/signup", controllers.RegisterHandler)
 
 		// The following routes will be authenticated
-		v1.Use(api.JWTAuthMiddleware())
+		v1.Use(service.JWTAuthMiddleware())
 
 		// route
 		route.FollowshipRouter(v1)
